pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "google-api-python-client>=2.177.0",
    "oauthlib>=3.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
    "yt-dlp>=2025.7.21",
]
